Split lambda into module specific files with exports and imports. One for EC2 methods, one for alarm tools and the handler remains the same. This is to establish a pattern for maintainability over the course of development across different AWS products and services that will be interacted with using AutoAlarm.
