+++ handlers/src/alarm-tools.ts
@@ -1,0 +1,237 @@
+import {
+  CloudWatchClient,
+  DeleteAlarmsCommand,
+  DescribeAlarmsCommand,
+  PutMetricAlarmCommand,
+} from '@aws-sdk/client-cloudwatch';
+import * as logging from '@nr1e/logging';
+import {AlarmProps, Tag} from './types';
+
+const log = logging.getRootLogger();
+const cloudWatchClient = new CloudWatchClient({});
+
+export async function doesAlarmExist(alarmName: string): Promise<boolean> {
+  const response = await cloudWatchClient.send(
+    new DescribeAlarmsCommand({AlarmNames: [alarmName]})
+  );
+  return (response.MetricAlarms?.length ?? 0) > 0;
+}
+
+export async function needsUpdate(
+  alarmName: string,
+  newProps: AlarmProps
+): Promise<boolean> {
+  try {
+    const existingAlarm = await cloudWatchClient.send(
+      new DescribeAlarmsCommand({AlarmNames: [alarmName]})
+    );
+
+    if (existingAlarm.MetricAlarms && existingAlarm.MetricAlarms.length > 0) {
+      const existingProps = existingAlarm.MetricAlarms[0];
+
+      if (
+        Number(existingProps.Threshold) !== newProps.threshold ||
+        Number(existingProps.EvaluationPeriods) !==
+          newProps.evaluationPeriods ||
+        Number(existingProps.Period) !== newProps.period
+      ) {
+        log.info().str('alarmName', alarmName).msg('Alarm needs update');
+        return true;
+      }
+    } else {
+      log.info().str('alarmName', alarmName).msg('Alarm does not exist');
+      return true;
+    }
+
+    log.info().str('alarmName', alarmName).msg('Alarm does not need update');
+    return false;
+  } catch (e) {
+    log.error().err(e).msg('Failed to determine if alarm needs update:');
+    return false;
+  }
+}
+
+export function configureAlarmPropsFromTags(
+  alarmProps: AlarmProps,
+  tags: Tag,
+  thresholdKey: string,
+  durationTimeKey: string,
+  durationPeriodsKey: string
+): void {
+  // Adjust threshold based on tags or use default if not present as defined in alarm props
+  if (!tags[thresholdKey]) {
+    log.info().msg('Threshold tag not found, using default');
+  } else if (tags[thresholdKey]) {
+    const parsedThreshold = parseFloat(tags[thresholdKey]);
+    if (!isNaN(parsedThreshold)) {
+      alarmProps.threshold = parsedThreshold;
+      log
+        .info()
+        .str('tag', thresholdKey)
+        .num('threshold', parsedThreshold)
+        .msg('Adjusted threshold based on tag');
+    } else {
+      log
+        .warn()
+        .str('tag', thresholdKey)
+        .str('value', tags[thresholdKey])
+        .msg('Invalid threshold value in tag, using default');
+    }
+    // Adjust period based on tags or use default if not present as defined in alarm props
+    if (!tags[durationTimeKey]) {
+      log.info().msg('Period tag not found, using default');
+    } else if (tags[durationTimeKey]) {
+      let parsedPeriod = parseInt(tags[durationTimeKey], 10);
+      if (!isNaN(parsedPeriod)) {
+        if (parsedPeriod < 10) {
+          parsedPeriod = 10;
+          log
+            .info()
+            .str('tag', durationTimeKey)
+            .num('period', parsedPeriod)
+            .msg(
+              'Period value less than 10 is not allowed, must be 10. Using default value of 10'
+            );
+        } else if (parsedPeriod < 30) {
+          parsedPeriod = 30;
+          log
+            .info()
+            .str('tag', durationTimeKey)
+            .num('period', parsedPeriod)
+            .msg(
+              'Period value less than 30 and not 10 is adjusted to 30. Using default value of 30'
+            );
+        } else {
+          parsedPeriod = Math.ceil(parsedPeriod / 60) * 60;
+          log
+            .info()
+            .str('tag', durationTimeKey)
+            .num('period', parsedPeriod)
+            .msg(
+              'Period value not 10 or 30 must be multiple of 60. Adjusted to nearest multiple of 60'
+            );
+        }
+        alarmProps.period = parsedPeriod;
+      } else {
+        log
+          .warn()
+          .str('tag', durationTimeKey)
+          .str('value', tags[durationTimeKey])
+          .msg('Invalid period value in tag, using default 60 seconds');
+      }
+    }
+    // Adjust evaluation periods based on tags or use default if not present as defined in alarm props
+    if (!tags[durationPeriodsKey]) {
+      log.info().msg('Evaluation periods tag not found, using default');
+    } else if (tags[durationPeriodsKey]) {
+      const parsedEvaluationPeriods = parseInt(tags[durationPeriodsKey], 10);
+      if (!isNaN(parsedEvaluationPeriods)) {
+        alarmProps.evaluationPeriods = parsedEvaluationPeriods;
+        log
+          .info()
+          .str('tag', durationPeriodsKey)
+          .num('evaluationPeriods', parsedEvaluationPeriods)
+          .msg('Adjusted evaluation periods based on tag');
+      } else {
+        log
+          .warn()
+          .str('tag', durationPeriodsKey)
+          .str('value', tags[durationPeriodsKey])
+          .msg('Invalid evaluation periods value in tag, using default 5');
+      }
+    }
+  }
+}
+
+export async function createOrUpdateAlarm(
+  alarmName: string,
+  instanceId: string,
+  props: AlarmProps,
+  tags: Tag,
+  thresholdKey: string,
+  durationTimeKey: string,
+  durationPeriodsKey: string
+) {
+  try {
+    log
+      .info()
+      .str('alarmName', alarmName)
+      .str('instanceId', instanceId)
+      .msg('Configuring alarm props from tags');
+    configureAlarmPropsFromTags(
+      props,
+      tags,
+      thresholdKey,
+      durationTimeKey,
+      durationPeriodsKey
+    );
+  } catch (e) {
+    log.error().err(e).msg('Error configuring alarm props from tags');
+    throw new Error('Error configuring alarm props from tags');
+  }
+  const alarmExists = await doesAlarmExist(alarmName);
+  if (!alarmExists || (alarmExists && (await needsUpdate(alarmName, props)))) {
+    try {
+      await cloudWatchClient.send(
+        new PutMetricAlarmCommand({
+          AlarmName: alarmName,
+          ComparisonOperator: 'GreaterThanThreshold',
+          EvaluationPeriods: props.evaluationPeriods,
+          MetricName: props.metricName,
+          Namespace: props.namespace,
+          Period: props.period,
+          Statistic: 'Average',
+          Threshold: props.threshold,
+          ActionsEnabled: false,
+          Dimensions: props.dimensions,
+        })
+      );
+      log
+        .info()
+        .str('alarmName', alarmName)
+        .str('instanceId', instanceId)
+        .num('threshold', props.threshold)
+        .num('period', props.period)
+        .num('evaluationPeriods', props.evaluationPeriods)
+        .msg(`${alarmName} Alarm configured or updated.`);
+    } catch (e) {
+      log
+        .error()
+        .err(e)
+        .str('alarmName', alarmName)
+        .str('instanceId', instanceId)
+        .msg(
+          `Failed to create or update ${alarmName} alarm due to an error ${e}`
+        );
+    }
+  }
+}
+
+export async function deleteAlarm(
+  instanceId: string,
+  check: string
+): Promise<void> {
+  const alarmName = `AutoAlarm-EC2-${instanceId}-${check}`;
+  const alarmExists = await doesAlarmExist(alarmName);
+  if (alarmExists) {
+    log
+      .info()
+      .str('alarmName', alarmName)
+      .str('instanceId', instanceId)
+      .msg('Attempting to delete alarm');
+    await cloudWatchClient.send(
+      new DeleteAlarmsCommand({AlarmNames: [alarmName]})
+    );
+    log
+      .info()
+      .str('alarmName', alarmName)
+      .str('instanceId', instanceId)
+      .msg('Deleted alarm');
+  } else {
+    log
+      .info()
+      .str('alarmName', alarmName)
+      .str('instanceId', instanceId)
+      .msg('Alarm does not exist for instance');
+  }
+}
